Removed Stackdriver Logging
StackDriver logging was added to provide structured logging. Apparently,
it wasn't really tested in the context of Google Cloud Functions. We
were experiencing a couple of issues:

* Random Crashing (#25)

* Missing Logs
  The function would sometimes exit without sending all logs.

* Missing Tags
  The previous logging implementation didn't provide tags with the
  function execution id. Standard output logs do this automatically.

While structured logging is great, the costs outweigh the benefits in
this case (unfortunately). So I'm switching back to stdout/stderr based
text logs.