Added TODO
The agency saving happens kinda implicitly. Due to the architecture of
dialogflow, there doesn't seem to be a simple way to send multiple
responses. The client drives the interaction instead of the server. The
client library doesn't buffer messages, but sends the response right
away. An abstraction between the generated response and the code
generating the responses could make this possible but would require a
large refactor.